src/goTest.ts: allow users to select default debug adapter
Add a configuration to delveConfig, to allow users to select which
debug adapter to use by default (codelenses included). The options
are the same as for launch.json.

Additionally, we take the directory containing the tests instead of
the test file, because 'go test ./file_test.go' does not include
the package which contains './file_test.go', so we should be passing
the package.

Fixes golang/vscode-go#1293

Change-Id: Ie0feb5fc2ba40532455726488e3857a8cff2e1a9
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/306969
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>