test/integration/goDebug: set mode to 'debug' or 'test' not 'auto'
When mode is 'auto', it is resolved to 'debug' or 'test' based on the
open file. However, we do not explicitly open any files, so we are
running these tests based on whatever the last opened file was. Instead
we want to explicitly set the mode to 'debug' or 'test'.

Change-Id: Ibe3a571c640292edb75434323623bfbebb18bfeb
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/312750
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>