test/integration/goDebug: verify goroutines running before switch
The test to make sure switch goroutines works by stepping between
two different goroutines. However, it was possible that when the
breakpoint in one goroutine was hit, the other had not started
running, so we were unable to get the threadID. This change sets
a breakpoint in each goroutine, to ensure that both are in the
appropriate functions before starting the tests.

Change-Id: I157a7d73a83722545db0878892e61079915043c8
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/313530
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>