test/integration/goDebug: don't wait for debug client to stop in teardown
We do not actually need to wait for the debug client to stop for the
test to be complete. The example code from testsupport calls dc.stop()
and does not wait for it to return. This should hopefully help with the
many timeouts we are experiencing.

Change-Id: I21425c26232e9553ea5b22937e3eb3cb109b5736
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/314429
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>