src/goDebugConfiguration: use fileWorkspaceFolder in multi root workspace
In multi folder workspaces, it is not allowed to use `${workspaceFolder}`.
If we detect that it is a multi folder workspace, we instead use
`${fileWorkspaceFolder}` which will get the folder of the current open
file.

Fixes golang/vscode-go#1499

Change-Id: Ibbe6e869cb0e5093c1bcd9f55d8f7057bceeed9f
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/319809
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>