src/goDebugConfiguration: resolve relative paths used in cwd, output, program
If relative paths are used, translate them to be relative to the workspace folder when
using dlv-dap.

The description in the package.json says cwd is a workspace relative or absolute path, but
this seems to be broken in the old adapter (I.e., when cwd=., the old adapter simply used it
as --wd value and launched the headless server in the program directory. As a result, '.'
is translated as the program or package source directory).
This CL doesn't attempt to fix or change the behavior of the old adapter though, but
applies the translation only when dlv-dap is used.

This changes the default cwd value (when users attempt to add cwd to their launch config)
to be '' which is treated as if 'cwd' attribute was undefined. Users who want to use the
workspace folder can use `${workspaceFolder}` or  `.`.

This change doesn't change 'cwd' in attach mode because this is currently used for
different purpose in the legacy adapter, and it will become irrelevant in dlv-dap.

Updates golang/vscode-go#1348

Change-Id: Ieb15f6bbb470a17d2e7350ccf1d8a003cbb92eeb
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/317210
Trust: Hyang-Ah Hana Kim <hyangah@gmail.com>
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Suzy Mueller <suzmue@golang.org>