Ensure that `cursorOrPrevious` always saves
In the case of `go.test.cursorOrPrevious` command not finding a Go test
under the cursor, it will call `go.test.previous`. However,
`cursorOrPrevious` automatically saves when there was a test found, but
`previous` does not. This can lead to the user being surprised as to why
doesn't `cursorOrPrevious` always save.