package.json: add config to hide system goroutines in debug
This change includes the configuration for hiding the system
goroutines in a debug session. It also sets the default value in
go nightly to true.

The setting of go.useLanguage server in nightly to true is also
removed by this change, since it is not the default everywhere.

Change-Id: I43b45b7743e6b3b31a41da1462e50669846b0f0d
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/359402
Trust: Suzy Mueller <suzmue@golang.org>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>