src/goEnv.ts: pass uri to getGoConfig
The newEnvironment function adds a parameter for uri
so that it can be passed to getGoConfig. This ensures
that the correct toolsEnvVars are provided to commands
that use toolExecutionEnvironment for env values.

Change-Id: I78bb430e8d5a198cefb46103a966f37daefba620
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/388434
Trust: Jamal Carvalho <jamal@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>