src/goDebugConfiguration: remove remote attach info pop-up
This change removes an informational pop-up at the start of remote attach session that was meant to announce to users that their dlv-dap setting was no longer ignored in that mode. In the past we defaulted to legacy regarding of the debugAdapter setting. Now that the intro period is over and the users also find this "announcement" confusing because it looks like an error, it is time to remove it. The users will be notified later in the session if their dlv version is too old to support remote attach and connection fails.

Fixes golang/vscode-go#2018

Change-Id: Iebecb3e7cb03d21a5a7b739c18554926a1e557f6
GitHub-Last-Rev: 3398531a05b4ef1ea42597a6866194c2286eb9b0
GitHub-Pull-Request: golang/vscode-go#2097
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/391054
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
Trust: Jamal Carvalho <jamal@golang.org>