src/util.ts: throw error when getGoVersion fails
In most cases the use of getGoVersion is expected to return a value.
Updating the function to throw an error in the case that a Go
version is not found makes this explicit. For the exceptions where
an explicit binary path is requested we handle the error thrown and
show the existing warnings.

For golang/vscode-go#57.

Change-Id: Ib80ceec0636134b6694b0080ff19e0295d49d59f
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/400996
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>