src/goDebugConfiguration: fix debugAdapter default selection
The intention of cl/399077 is to choose 'legacy' as the default
for 'remote' mode when we are not in preview mode, while choose
'dlv-dap' when we are in preview mode. The condition was reversed.
CI couldn't detect the bug because it was running only in
preview mode.

For golang/vscode-go#2168
Fixes golang/vscode-go#2215

Change-Id: I31b353abce29e52270cc589916c61aa5712258e3
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/402220
Reviewed-by: Polina Sokolova <polina@google.com>