src/goVulncheck: add VulncheckResultViewProvider
VulncheckResultViewProvider provides a custom text document editor
to process `gopls vulncheck` JSON output and render it in a Webview.
The webview presents the data in a similar way `govulncheck -html`
presents the summary of the findings. But by implementing it inside
the extension, we can make the extension handle commands and document
links inside the editor. For example, in this CL, we embed a js script
that listens 'click' event on all links in the HTML document, and
delegates the extension part to handle link opening events. That way,
if the resource is a file type, the extension can open it inside the
editor.

For communication between the Extension side and the WebView side,
we use vscode's API to pass messages and follow the instruction in
https://code.visualstudio.com/api/extension-guides/webview#scripts-and-message-passing

The extension sends 'update' msg to the webview: that triggers update
of DOM contents based on the input.

The webview sends 'link' msg to the webview: that triggers handling
of the link opening event. All file links in the callgraph entries
carry the position information as a query parameter. The extension
parses the uri, open the file with vscode.openWith command (that allows
us to control which view column the editor should be placed, and which
line in the file should get focus).

In order to support testing, the extension and the webview exchange
'snapshot-request' and 'snapshot-result' messages. When we have UI testing
framework established, these message may be able to be replaced with it.

Change-Id: I9a343b972642a9197313d9d42af37b5a2d5ccefa
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/406300
TryBot-Result: kokoro <noreply+kokoro@google.com>
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>