Open file in corresponding workspace
It took me some time to understand why autocomplete does not work properly.

See https://stackoverflow.com/a/72743549/633961

I guess you want a different wording. Feel free to change the text of my PR. But I think
a hint like this needs to be at the top of the file.