test/integration/goTest: fix populateModulePathCache
And reenable affected tests.

The tests were broken due to a bug in composing the file Uris for
files in a directory from the directory Uri. path.join uses the
platform specific separator (\ on windows). To ensure the correct
encoding, convert the directory Uri to its fsPath, construct the
file path of the file, and convert the file path to Uri.

For golang/vscode-go#832
For golang/vscode-go#1788

Change-Id: I3dfc44e04f1683b1b862b339b7257b432c3f2241