test/integration/goDebug: make invalid flag test faster
This test is to check if the debugging session fails
somehow if an invalid dlv flag is passed.
Previously, this test relied on the timeout of dap server
start. (See spawnDlvDapServerProcess and its use of
30_000msec timeout) That was why this test always took
30sec to complete.

This CL changes the test to wait for the output message
that includes the usage error information, instead of
waiting for the timeout.

I think we can also improve spawnDlvDapServerProcess
to fail sooner when dlv server start up failure is observed
instead of relaying on the timeout. However, the start up
logic is complex enough. I am reluctant to add more
complexity to optimize for this less common failure case.

Change-Id: I16069dc1c6879ce137b7799366e36e429500ea53
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/418899
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>