src/goTest: open debug repl when running debug test
This change focuses on the debug console when the debug test is run.
This does not guarantee that the active debug console is for the
debug test session if there are multiple debug sessions running.

Fixes golang/vscode-go#2409

Change-Id: I5ac77ab9a1d5de90bc52c148e751159c3b7a9457
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/424930
TryBot-Result: kokoro <noreply+kokoro@google.com>
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>