src/goInstallTools: handle unknown tools
For example, when the extension detects that the specified custom formatting
tool does not exist, it will attempt to install or prompt the user for installing
the missing tool. However, it's likely the tool is not known to our extension,
so getTool won't be able to return a valid Tool. Handle this case by checking
the return value of getTool.

I expect there are still many corner cases and UX bugs in handling custom
formatter errors. But, assuming use of custom formatters is rare, polishing
the custom formatter UX is a low priority.

For golang/vscode-go#1238

Change-Id: If6deedf9285c6d070ad4d960f78c5430be54afd5
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/446299
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>