goplsSurvey: remove lumpiness in survey days
Users are occasionally asked to take the gopls survey. The logic
is a little complicated, but in the existing code, if the user is
prompted, it will be for a random day in the rest of the current
month. In the new code all days of the year are pretty nearly equally
probable. The new code chooses a random day in the next 28 days.

There's nothing special about 28, but it's large enough to smooth out
day-of-the-week effects, and small enough that a chosen user
will be prompted for a survey within a month.

Fixes vscode-go/#2545

Change-Id: Ifaccaa5e8e4e8b1a1050768461ed38d4e144561e
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/456041
TryBot-Result: kokoro <noreply+kokoro@google.com>
Run-TryBot: Peter Weinberger <pjw@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>