src/language/goLanguageServer: support staged auto-update of gopls
vscode-go periodically runs go list to check for the new gopls release
and, if auto update is enabled, it updates the gopls as soon as it
detects the new version. We do not know how many users have enabled
this auto-update feature nor how soon they detect the new release
(this depends on when the user started the code, etc) without telemetry.
But we think this can be too risky.

This change implements staged rollout for auto-update - when a
new non-patch version is released, the extension checks the version's
timestamp. If the release is less than a day old, update with probability
0.1; less than 3days old, update with probability 0.3; update if older.
The probability is deterministically computed based on the machine id
and the gopls version number.

For security releases or fixes to address regression, we will need to
upgrade faster. We assume those are released as patch versions and update
immediately.

The extension picks pre-release versions only if the extension is in
preview mode (e.g. Nightly or dev version). Thus, we proceed to upgrade
immediately if the detected and chosen version is a pre-release.

Change-Id: Ib12b455c65b3516e81dfcc10d7efa05f14c4af35
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/468497
Reviewed-by: Robert Findley <rfindley@google.com>
Run-TryBot: Hyang-Ah Hana Kim <hyangah@gmail.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>