src/pickProcess: show error when the process picker is hidden
If a user accidentally hides the UI for the process picker, the
debugging setup process will hang, and it is unclear how to stop
or restart the process. This makes sure to dispose of the ui
element when we are finished with it and rejects the input if the
ui is hidden.

Change-Id: I32516ffbde7f2a48c59439769b44806c969a642e
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/499601
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>