build: add GCB release workflow
This workflow is meant to replace the GitHub action workflow defined in
https://github.com/golang/vscode-go/blob/master/.github/workflows/release.yml

The GH workflow was a collection of command line tools invocation and
some bash scripting to parse the version tag and adjust the workflow
depending on whether it's for rc or not. I figured it's easier to work
with go, so ported the logic to build/release.go.

TODO: convert most part of build/all.bash and the nightly release workflow
with go code.

release.go uses `gh` (GitHub CLI tool) to post the release note and
the artifacts (vsix) to GitHub. Dockerfile is updated to include `gh`.
Reference: https://stackoverflow.com/a/69477930

The workflow needs to access GitHub personal access token.
To access the release API, write permission for Content & Workflow is
necessary.
https://docs.github.com/en/rest/authentication/permissions-required-for-fine-grained-personal-access-tokens?apiVersion=2022-11-28#repository-permissions-for-contents

Tested it by triggering the GCB workflow against a forked repo.

For b/216321767

Change-Id: I1eec6b4b896fe1f1565bbf9e160f0368cade36c7
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/553115
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Suzy Mueller <suzmue@golang.org>
Commit-Queue: Hyang-Ah Hana Kim <hyangah@gmail.com>
TryBot-Result: kokoro <noreply+kokoro@google.com>