extension/src/welcome: readd the fix for go.showWelcome handling
The original fix for golang/vscode-go#3319 also included
a minor refactoring attempt - centralize welcome page show logic
in shouldShowGoWelcomePage. But that unfortunately introduced
a new bug, i.e., failed to apply the Web-based IDE exclusion rule
correctly.

Reverted the previous change to remove the refactoring.
And, this change adds the specific fix to address the bug reported
in golang/vscode-go#3319.

Test coverage would be nice, but it is tricky without
major refactoring or complicating test logic to stub or
inject dependencies.

Renamed shouldShowGoWelcomePage to make it clear that
this is to decide whether we have news to show purely
based on the extension's version.

For golang/vscode-go#3319.

Change-Id: Ia4915e0201e73d136b3efcec7c0cf4f5a5e4559f
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/581118
kokoro-CI: kokoro <noreply+kokoro@google.com>
Commit-Queue: Hyang-Ah Hana Kim <hyangah@gmail.com>
Reviewed-by: Peter Weinberger <pjw@google.com>