extension/src/goSurvey: fix a bug in HaTS survey prompting
Once a month, the extension picks the date to prompt if the
user is chosen for HaTS survery. The extension then arranges
setTimeout to prompt after the picked date.
The picked date is stored in `dateToPromptThisMonth` (memento)
so we can continue the timer even when vscode restarts.

However, there was a bug that miscompute the wait time for
setTimeout. If `dateToPromptThisMonth` has already passed,
it waited for now-dateToPromptThisMonth, which was unintended
and unnecessary. Fix it.

It's posible that this can increase # of HaTS survey
responders slightly.

Change-Id: I305c1070b673915426bcd619642df13b6aea66ce
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/593821
Auto-Submit: Hyang-Ah Hana Kim <hyangah@gmail.com>
Commit-Queue: Hyang-Ah Hana Kim <hyangah@gmail.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>