extension: stop setting "preview" for v0.43.0
v0.43.0 is our first pre-release version that replaces
the nightly extension. Pre-release is built from the master
branch. Previously we set the preview' field because
the nightly extension is built from the master and
golang.go-nightly is a preview extension in our
definition. golang.go is not a preview extension
so let's unset this field.

Nightly extension publishing is paused.

And add a CHANGELOG entry for v0.43.0.
For pre-release versions, we will skip CHANGELOG addition.
But, it's good to announce the first pre-release version
and mention the deprecation of nightly.

For golang/vscode-go#1935

Change-Id: I326963818aa3bd7723bbdf69d5b747f57210b5c1
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/607756
Reviewed-by: Robert Findley <rfindley@google.com>
Commit-Queue: Hyang-Ah Hana Kim <hyangah@gmail.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>