extension: update go to go1.23.1
We want to use the latest go when developing our own
release/testing support tools and scripts.
So, let's update go.mod to do so.

After go1.21, we shouldn't use `go run` to run
the extension/tools/installtools script.
The purpose of `installtools` is to install tools
needed by integration tests and are compatible with
the go version in the system. We arranged the CI
systems to test with the go versions we support.
If we use `go run` from our project, the toolchain
switch will occur and may change `GOROOT`/`PATH`
to enforce the compiled script to use the upgraded
toolchain. That makes the CI setup useless.

Therefore, in this CL, we build the binary (it's ok to
build the binary with go1.23.1+), and then run
the installed binary ourselves so the execution of
the binary doesn't get affected by the modified
GOROOT/PATH.

Kokoro CI also uses the installtools script when
building the docker container. (build/Dockerfile)
There, the script source code is copied over to a
scratch space and run with `go run <go file>`
outside the vscode-go project repo. So, there is
no go toolchain version switch involved already.

And currently we test only with the latest go in
Kokoro, so this toolchain switch issue doesn't apply.

For golang/vscode-go#3411

Change-Id: I3e116cf48fb431196359ec42049e70c0b75814ef
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/616677
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hongxiang Jiang <hxjiang@golang.org>
Commit-Queue: Hyang-Ah Hana Kim <hyangah@gmail.com>