extension/src/goDebugConfiguration: fix debugging package urls
A package URL such as example.com/foo/bar is a valid program for dlv
debug. This updates goDebugConfiguration's handling of the program
attribute to prevent it from A) treating it as a path (prepending the
workspace folder path) and B) throwing an error due to it not being a
valid filesystem path.

Updates golang/vscode-go#1627.

Change-Id: I9ad15752271c84fa069106c3a96206b6edcc3797
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/661195
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Hongxiang Jiang <hxjiang@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>