extension/src/goLint: adding `--path-mode=abs` flag for golangci-lint v2
This change sets path mode to abs (supported from golangci-lint v2.1) when running golangci-lint v2.
When `run.relative-path-mode` is set to `cfg` (which is the default), the file path in the linter output is relative to the config file, while vscode-go treats it as relative to the linted file. That makes the file path invalid, which prevents diagnostic errors from showing.
The absolute output path is the efficient way to fix this problem.

Fixes golang/vscode-go#3750

Change-Id: I77bc8a54062bad094128a14d764db94198379ee9
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/664877
Reviewed-by: Hongxiang Jiang <hxjiang@golang.org>
Reviewed-by: Madeline Kalil <mkalil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
kokoro-CI: kokoro <noreply+kokoro@google.com>