[release-v0.48]extension/test/integration: update debug config for remote mode
CL 643280 changed the behavior of debug adapter in remote mode from
always "legacy" to the result from command
`dlv substitute-path-guess-helper`.

This CL calls the method to find the right return value and test the
debug adapter with the command return.

In master branch, the integration test is running against preview
version of extension, so the debug adapter is always "dlv-dap"
regardless of the command return. So this test is not captured
in the master branch.

Change-Id: I5ce6e1040fa16f10b038265d7e0ab7a884a99141
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/684475
kokoro-CI: kokoro <noreply+kokoro@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Madeline Kalil <mkalil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>