extension/src: use full path mode when running go test
The fullpath mode is added in go 1.21 and the vscode-go is supporting
the last three minor Go versions (1.23.0+ as of 2025-09).

If the user is beyond the support, the extension will send a
notification asking user to upgrade.

In case user does not want to upgrade, set "-test.fullpath=false"
to setting "go.testFlags" will fall back to the original behavior.

Fix golang/vscode-go#3853

Change-Id: I06b723b715651ea0bf98ae62797917af7b1c74b2
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/704776
Reviewed-by: Madeline Kalil <mkalil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>