extension/src: remove default setting for "go.lintTool"
The "lintTool" will be the source of truth indicating which tool
will be run as linter. vscode-go will now execute the exactly
tool based on this setting without any hidden yield logic.

Previously, the default value for "go.lintTool" is "staticcheck",
vscode-go have special logic that yield the linter functionality
to gopls if the lint tool is staticcheck and gopls staticheck is
enabled. It confuses the user which tool will be executed.

This CL removes this default option from "go.lintTool", consider
the new default as "". An empty lintTool indicates the user does
not want any linter to run from the extension.

This CL also change the description of "go.lintTool" to the lint
tool in addition to gopls, also mention in the description that
the linting functionality provided by gopls (the language server)
can be configured in a seprate setting.

To eliminate further confusions, a validation logic will be
triggered when gopls or go configuration changed: if "staticcheck"
is chosen as lintTool and gopls staticcheck is enabled. Warning
will be sent to user to disable one of them to avoid duplicate
diagnostics.

Alternative considered: introduce "gopls" as default lintTool.

For golang/vscode-go#3862

Change-Id: Ia00ca115ca7ec34315043ad61517b5cdfb9ca1ed
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/709035
Reviewed-by: Madeline Kalil <mkalil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>